requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.0
pillow>=8.2.0
requests>=2.25.1
pathlib>=1.0.1
scikit-learn>=1.0.0
tensorflowjs>=3.12.0
